Authr

+++ TakeASeat/Controllers/SeatsReservationController.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Data;
 using TakeASeat.Data;
 using TakeASeat.Data.DatabaseContext;
 using TakeASeat.RequestUtils;
@@ -30,28 +32,37 @@ namespace TakeASeat.Controllers
             _context = serviceProvider.GetRequiredService<DatabaseContext>();
             
         }
-        [HttpGet]
-        public async Task<IActionResult> test()
-        {
-            await _releaseResRepository.ReleaseUnpaidReservations();
-            return Ok();
-        }
+        
 
         [HttpPost("create")]
+        [ApiVersion("1.0")]
+        [Authorize(Roles = "Administrator,Organizer,User")]
         [ProducesResponseType(StatusCodes.Status202Accepted)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> ReserveSeats([FromBody] RequestOrderParams rParams)
         {
-
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
             var seats = _mapper.Map<IList<Seat>>(rParams.Seats);
             await _seatResRepository.CreateSeatReservation(rParams.UserId, seats);
             return StatusCode(202);
         }
 
         [HttpPost("delete")]
+        [ApiVersion("1.0")]
+        [Authorize(Roles = "Administrator,Organizer,User")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> deleteSeatReservation([FromBody] RequestReservationParams seatReservation)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
 
             await _seatResRepository.DeleteSeatReservation(seatReservation.seatReservationId);
